modules/context/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio>=4.0.0
